# infrastructure/api/openapi.yaml
openapi: 3.0.2
info:
  title: API-First Spring Boot Example
  description: "Simple example of API-First integration in a Spring Boot application!"
  version: 1.0.0-SNAPSHOT
...
paths:
...
  /auth/refreshtoken:
    post:
      tags:
        - auth
      summary: log in user
      description: generate JWT
      operationId: refreshToken
      requestBody:
        description: Retrieve token for user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshAuthRequest'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
  /auth/registration:
    post:
      tags:
        - auth
      summary: register a new user as customer user
      description: generate JWT
      operationId: registration
      requestBody:
        description: Register a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Email already used
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDTO'
...
components:
  schemas:
...
    AuthRequest:
      type: object
      required:
        - password
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
    RefreshAuthRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        token:
          type: string
        refreshToken:
          type: string
...
    ErrorDTO:
      type: object
      properties:
        code:
          type: string
          example: PET_NOT_FOUND
        error:
          type: string
          example: The pet was not found